Quadtree, Clock, and README update
Implemented the quadtree to the separation. cohesion, and alignment of the boids. It runs at a steady 80fps with 200 boids currently.

Implemented a clock that decouples the movement of the boids from the framerate. They will now continue to move smoothly even when the framerate is choppy and won't move quickly if the framerate is high.

Updated the README. It now includes a short description of boids. I have also added instructions to install g++, and updated the "What's coming?" section: renaming it to "Future Updates" and removing the sections for added updates. I've edited the "Performance Overall" section to be called "Performance Increase" and reflect the future plans for the quadtree (and adding its insert method).